move WatchService to service package
