+++ android/app/src/main/AndroidManifest.xml
@@ -68,7 +68,7 @@
             android:name="flutterEmbedding"
             android:value="2" />
 
-        <service android:name=".WatchService" />
+        <service android:name=".service.WatchService" />
         <service
             android:name=".notifications.NotificationListener"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE">
