+++ android/app/src/main/kotlin/io/rebble/fossil/service/WatchService.kt
@@ -1,4 +1,4 @@
-package io.rebble.fossil
+package io.rebble.fossil.service
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -10,6 +10,8 @@ import androidx.annotation.DrawableRes
 import androidx.core.app.NotificationCompat
 import androidx.lifecycle.LifecycleService
 import androidx.lifecycle.lifecycleScope
+import io.rebble.fossil.FossilApplication
+import io.rebble.fossil.R
 import io.rebble.fossil.bluetooth.ConnectionLooper
 import io.rebble.fossil.bluetooth.ConnectionState
 import io.rebble.libpebblecommon.ProtocolHandler
